requirements.txt
anyascii==0.3.1
apimatic-core==0.1.2
apimatic-core-interfaces==0.1.0
apimatic-requests-client-adapter==0.1.0
asgiref==3.5.2
autopep8==2.0.0
Babel==2.11.0
beautifulsoup4==4.11.1
CacheControl==0.12.11
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
deprecation==2.1.0
Django==4.0.8
django-allauth==0.51.0
django-extra-views==0.13.0
django-filter==22.1
django-haystack==3.2.1
django-modelcluster==6.0
django-oscar==3.1
django-oscar-api==3.1.2
django-oscar-api-checkout==3.1.0
django-permissionedforms==0.1
django-phonenumber-field==5.2.0
django-rest-auth==0.9.5
django-salesman==1.1.4
django-tables2==2.3.4
django-taggit==3.0.0
django-treebeard==4.4
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
draftjs-exporter==2.1.7
enum34==1.1.10
et-xmlfile==1.1.0
factory-boy==3.2.1
Faker==15.1.3
html5lib==1.1
idna==3.4
importlib-metadata==5.0.0
jsonpickle==1.5.2
l18n==2021.3
Markdown==3.4.1
msgpack==1.0.4
oauthlib==3.2.2
openpyxl==3.0.10
packaging==21.3
phonenumbers==8.12.57
Pillow==9.2.0
psycopg2==2.9.5
purl==1.6
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sorl-thumbnail==12.9.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
squareup==21.1.0.20220823
tablib==3.2.1
telepath==0.3
tomli==2.0.1
tzdata==2022.5
urllib3==1.26.12
wagtail==4.0.4
webencodings==0.5.1
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
zipp==3.10.0
